+++ packages/gasket-cli/src/config/default-plugins.js
@@ -3,5 +3,6 @@ module.exports = [
   require('@gasket/plugin-lifecycle'),
   require('@gasket/plugin-metadata'),
   require('@gasket/plugin-start'),
-  require('@gasket/plugin-git')
+  require('@gasket/plugin-git'),
+  require('@gasket/plugin-logger')
 ];