+++ packages/gasket-plugin-express/test/plugin.test.js
@@ -211,7 +211,7 @@ describe('createServers', () => {
 
   it('adds middleware from lifecycle (ignores falsy)', async () => {
     await plugin.hooks.createServers(gasket, {});
-    expect(app.use).toHaveBeenCalledTimes(2);
+    expect(app.use).toHaveBeenCalledTimes(3);
 
     app.use.mockClear();
     mockMwPlugins = [
@@ -220,7 +220,7 @@ describe('createServers', () => {
     ];
 
     await plugin.hooks.createServers(gasket, {});
-    expect(app.use).toHaveBeenCalledTimes(3);
+    expect(app.use).toHaveBeenCalledTimes(4);
   });
 
   it('supports async middleware hooks', async () => {
@@ -256,16 +256,35 @@ describe('createServers', () => {
     ];
     await plugin.hooks.createServers(gasket, {});
 
-    expect(app.use.mock.calls[2]).toContain(paths);
+    expect(app.use.mock.calls[3]).toContain(paths);
   });
 
   it('adds errorMiddleware from lifecycle (ignores falsy)', async () => {
     await plugin.hooks.createServers(gasket, {});
-    expect(app.use).toHaveBeenCalledTimes(2);
+    expect(app.use).toHaveBeenCalledTimes(3);
     lifecycles.errorMiddleware.mockResolvedValue([() => {}, null]);
     app.use.mockClear();
     await plugin.hooks.createServers(gasket, {});
-    expect(app.use).toHaveBeenCalledTimes(3);
+    expect(app.use).toHaveBeenCalledTimes(4);
+  });
+
+  it('attaches a logger to the request', async () => {
+    const augmentedLogger = { info: jest.fn() };
+    gasket.logger = { child: jest.fn().mockReturnValue(augmentedLogger) };
+    await plugin.hooks.createServers(gasket, {});
+
+    const req = {};
+    const res = {};
+    const next = jest.fn();
+
+    app.use.mock.calls.forEach(([middleware]) => middleware(req, res, next));
+    expect(req).toHaveProperty('logger', gasket.logger);
+
+    req.logger.metadata({ userId: '123' });
+    expect(gasket.logger.child).toHaveBeenCalledWith({ userId: '123' });
+
+    req.logger.info('test');
+    expect(augmentedLogger.info).toHaveBeenCalledWith('test');
   });
 
   function findCall(aSpy, aPredicate) {