+++ packages/gasket-plugin-redux/CHANGELOG.md
@@ -1,5 +1,8 @@
 # `@gasket/plugin-redux`
 
+- Use per-request logger
+- Remove dependency on `@gasket/plugin-log`
+
 ### 6.34.4
 
 - Upgrade eslint-plugin-unicorn v43 ([#436])